refactor(voct): extract AUX frequency measurement out of clock.rs
