+++ faderpunk/src/tasks/mod.rs
@@ -10,3 +10,4 @@ pub mod leds;
 pub mod max;
 pub mod midi;
 pub mod transport;
+pub mod voct_freq;
